skills/charpup/cisco-security-auditor/requirements.txt
# Phase 2 Dependencies
yara-python>=4.3.0
requests>=2.28.0
pyyaml>=6.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
